+++ Changelog.md
@@ -3,7 +3,7 @@
 - Rename playlists with the new button
 - Better app icon & notification icons
 - Better notification progress bar
-- Folders tab is now works
+- Folders, artists, albums, genres, composer tabs are now works
 - UI improvements
 - Lot of bug fixes & improvements
   - Fix duplicate query causes infinity loop